rr_graph: Move utility funcs to __init__.py
